+++ packages/root/src/types/location.ts
@@ -1,23 +1,18 @@
 export interface LocationDetails {
-  query?: any;
-  hash?: string;
-  state?: any;
-}
-
-export interface PartialLocation extends LocationDetails {
-  pathname?: string;
-  rawPathname?: string;
-}
-
-// use HickoryLocation instead of Location to prevent
-// errors from colling with window.Location interface
-export interface HickoryLocation {
   pathname: string;
   query: any;
   hash: string;
   state?: any;
+}
+
+export type PartialLocation = Partial<LocationDetails>;
+
+// use HickoryLocation instead of Location to prevent
+// errors from colliding with window.Location interface
+export interface HickoryLocation extends LocationDetails {
   key: string | undefined;
   rawPathname: string;
+  url: string;
 }
 
 export type AnyLocation = HickoryLocation | PartialLocation;